requirements.txt
fastmcp>=0.1.0
pyyaml>=6.0
Pillow>=10.0
openai>=1.0
python-dotenv>=1.0
